+++ cx-portal/src/components/pages/Home/index.tsx
@@ -18,7 +18,6 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import SearchSection from './components/SearchSection'
 import SearchResultSection from './components/SearchResultSection'
 import NewsSection from './components/NewsSection'
 import BusinessApplicationsSection from './components/BusinessApplicationsSection'
@@ -27,12 +26,13 @@ import AppStoreSection from './components/AppStoreSection'
 import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
 import './Home.scss'
+import NewSearchSection from './components/SearchSection/NewSearchSection'
 
 export default function Home() {
   return (
     <main className="home">
+      <NewSearchSection />
       <StageSection />
-      <SearchSection />
       <SearchResultSection />
       <NewsSection />
       <BusinessApplicationsSection />